crypto-ai-bot/requirements.txt
aiohttp==3.9.5
python-telegram-bot==21.3
python-dotenv==1.0.1
google-generativeai==0.7.2
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
feedparser==6.0.11
aiofiles==23.2.1
tenacity==8.3.0
loguru==0.7.2
matplotlib==3.9.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
pytz==2024.1
